+++ report/renderer/pwa-category-renderer.js
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {ReportUtils} from './report-utils.js';
+import {ReportUtils, SharedUtils} from './report-utils.js';
 import {CategoryRenderer} from './category-renderer.js';
 
 export class PwaCategoryRenderer extends CategoryRenderer {
@@ -101,7 +101,7 @@ export class PwaCategoryRenderer extends CategoryRenderer {
 
     // Remove any that have a failing audit.
     for (const auditRef of auditRefs) {
-      if (!ReportUtils.showAsPassed(auditRef.result) && auditRef.group) {
+      if (!SharedUtils.showAsPassed(auditRef.result) && auditRef.group) {
         uniqueGroupIds.delete(auditRef.group);
       }
     }
@@ -122,7 +122,7 @@ export class PwaCategoryRenderer extends CategoryRenderer {
     for (const groupId of groupIds) {
       const groupAuditRefs = auditRefs.filter(ref => ref.group === groupId);
       const auditCount = groupAuditRefs.length;
-      const passedCount = groupAuditRefs.filter(ref => ReportUtils.showAsPassed(ref.result)).length;
+      const passedCount = groupAuditRefs.filter(ref => SharedUtils.showAsPassed(ref.result)).length;
 
       const title = groupDefinitions[groupId].title;
       tips.push(`${title}: ${passedCount}/${auditCount}`);
